Fix: Extra navigation options not visible on login
Links to the search and pantry pages should appear immediately when logging in, but due to the order of the code these links were not shown until navigating to any page (other than the logout page) after logging in.

This fixes that problem, so that immediately after logging in all of the navigation links are displayed.