Check if options are set for pattern match recipe search
Also remove unnecessary default option (was set to false, now removed entirely).

I had mistakenly assumed that the elements in the default array would remain the default if not defined in an array passed to a method. In retrospect, it's obvous this wouldn't be the case. The method now explicitly checks if elements in the options argument are set.