Add search option to filter recipes by ingredients in pantry
This new option allows the search to filter not only by a search string, but also by whether the logged-in user has all the ingredients.

This should make searchByDBPantry redundant and irrelevant, because it is far more powerful.

I don't like that the method accesses $_SESSION. It works, but I would prefer if the user id was passed into the method. Still trying to think of a way to do that which isn't obtuse and confusing wrt variable names. "in_pantry" being set to true makes sense, ish, but setting it to the user id does not make sense. Having an option that takes the user id would make more sense, but only if the person using the method understood what the method was doing (which shouldn't be necessary). Currently... it works.